api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
spotipy==2.26.0
supabase==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic-settings==2.2.1
httpx==0.27.0
